Move config to utils package
