+++ cmd/utils/config.go
@@ -1,4 +1,4 @@
-package config
+package utils
 
 import (
 	"github.com/kelseyhightower/envconfig"
